re_web_viewer_server no longer needs tokio, split out sync code path
