+++ crates/re_sdk/Cargo.toml
@@ -68,7 +68,9 @@ thiserror.workspace = true
 re_data_source = { workspace = true, optional = true }
 re_smart_channel = { workspace = true, optional = true }
 re_ws_comms = { workspace = true, optional = true }
-re_web_viewer_server = { workspace = true, optional = true }
+re_web_viewer_server = { workspace = true, optional = true, features = [
+  "sync",
+] }
 
 anyhow = { workspace = true, optional = true }
 webbrowser = { workspace = true, optional = true }
